tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_display-inline.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | display: inline-flex</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#flex-containers" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	display: inline-flex;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("display"),
	          "inline-flex");
});
</script>
</body></html>
